update: replace wallaby with hyperspace
